Switch to reverse proxy instead of WASM for development
