+++ go/wallet/server.go
@@ -1,10 +1,15 @@
 package wallet
 
 import (
+	"context"
 	"embed"
+	"fmt"
 	"net/http"
+	"os"
+	"os/signal"
+	"syscall"
 
-	"github.com/gorilla/mux"
+	"github.com/onflow/fcl-dev-wallet/go/wallet/app"
 )
 
 //go:embed bundle.zip
@@ -14,23 +19,43 @@ var bundle embed.FS
 var envConfig []byte
 
 const bundleZip = "bundle.zip"
-
-type FlowConfig struct {
-	Address    string `json:"flowAccountAddress"`
-	PrivateKey string `json:"flowAccountPrivateKey"`
-	PublicKey  string `json:"flowAccountPublicKey"`
-	AccessNode string `json:"flowAccessNode"`
-	AvatarUrl string `json:"flowAvatarUrl"`
+type server struct {
+	http *http.Server
+	app *app.App
 }
 
-func buildHandlers(srv *server) http.Handler {
-	r := mux.NewRouter()
+// NewHTTPServer returns a new wallet server listening on provided port number.
+func NewHTTPServer(port uint, config *app.FlowConfig) (*server, error) {
+	app := app.NewApp(config, bundle, bundleZip, envConfig)
+	http := http.Server{
+		Addr: fmt.Sprintf(":%d", port),
+		Handler: app.Handler,
+	}
 
-	r.HandleFunc("/api/", configHandler(srv))
-	r.HandleFunc("/api/polling-session", pollingSessionHandler(srv))
-	r.HandleFunc("/api/{service}", serviceHandler(srv))
-	r.HandleFunc("/", devWalletHandler())
+	srv := &server{
+		http: &http,
+		app: app,
+	}
 
-	return r
+	return srv, nil
 }
 
+func (s *server) Start() {
+	done := make(chan os.Signal, 1)
+	signal.Notify(done, os.Interrupt, syscall.SIGINT, syscall.SIGTERM)
+
+	go func() {
+		err := s.http.ListenAndServe()
+		if err != nil {
+			fmt.Printf("error starting up the server: %s\n", err)
+			done <- syscall.SIGTERM
+		}
+	}()
+
+	<-done
+	s.Stop()
+}
+
+func (s *server) Stop() {
+	_ = s.http.Shutdown(context.Background())
+}
