+++ package.json
@@ -4,9 +4,9 @@
   "private": true,
   "license": "Apache-2.0",
   "scripts": {
-    "dev": "rimraf ./.polling-sessions.tmp && npm run build-wasm && APP_ENV=local next dev --port 8701",
+    "dev": "concurrently \"APP_ENV=local next dev --port 8701\" \"go run ./go/wallet/cmd/main.go --port 8799\"",
     "build": "next build",
-    "start": "next start",
+    "start": "concurrently \"next start\" \"go run ./go/wallet/cmd/main.go --port 8799\"",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "tsc": "./node_modules/typescript/bin/tsc",
@@ -16,8 +16,7 @@
     "bundle": "next build && next export",
     "zip": "cd ./out && zip -FSr ../go/wallet/bundle.zip . && cd ..",
     "config": "cp .env.development ./go/wallet",
-    "go-server": "npm run bundle && npm run zip && npm run config && go run ./go/wallet/cmd/main.go",
-    "build-wasm": "GOOS=js GOARCH=wasm go build -o ./api.wasm ./go/wallet/cmd/main_wasm.go"
+    "go-server": "npm run bundle && npm run zip && npm run config && go run ./go/wallet/cmd/main.go"
   },
   "dependencies": {
     "@emotion/react": "^11.9.0",
@@ -55,6 +54,7 @@
     "@typescript-eslint/parser": "^4.28.1",
     "babel-eslint": "^10.1.0",
     "babel-plugin-react-remove-properties": "^0.3.0",
+    "concurrently": "^8.2.0",
     "cypress": "^8.0.0",
     "cypress-iframe": "^1.0.1",
     "eslint": "^7.32.0",
